requirements.txt
Django==3.2
celery==5.2.7
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.21.3
flower==1.2.0
psycopg2-binary==2.9.3
python-dotenv~=0.21.0
requests==2.28.1
